refactor: 실제 응답 형식에 맞게 수정
